Rename HTML elements component, make minor code cleanup edits. Co-authored by: Janica Abagat abagat07@gmail.com Trisha Duong trishanduong@gmail.com David Lee davidetlee@gmail.com
