+++ src/components/right-sidebar/HTMLElementsList.vue
